sentinel_integrations/requirements.txt
openai
simple-salesforce
jinja2
fastapi
nats-py
pydantic-settings
